requirements.txt
accelerate==0.25.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bertopic==0.16.0
bleach==6.1.0
blis==0.7.11
bokeh==3.4.1
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorcet==3.1.0
colored==2.2.4
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
cucim-cu12==24.4.0
cuda-python==12.4.0
cudf-cu12==24.4.0
cugraph-cu12==24.4.0
cuml-cu12==24.4.0
cuproj-cu12==24.4.0
cupy-cuda12x==13.0.0
cuspatial-cu12==24.4.0
cuvs-cu12==24.4.0
cuxfilter-cu12==24.4.1
cycler==0.12.1
cymem==2.0.8
Cython==0.29.37
dask==2024.1.1
dask-cuda==24.4.0
dask-cudf-cu12==24.4.0
dask-expr==0.4.0
datasets==2.18.0
datashader==0.16.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.15.0
dill==0.3.8
distributed==2024.1.1
evaluate==0.4.1
executing==2.0.1
fastjsonschema==2.19.1
fastrlock==0.8.2
filelock==3.13.4
fiona==1.9.6
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
geopandas==0.14.3
h5py==3.11.0
hdbscan==0.8.33
holoviews==1.18.3
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
imageio==2.34.0
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
isoduration==20.11.0
janus==1.0.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_proxy==4.1.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
keybert==0.8.4
kiwisolver==1.4.5
langcodes==3.3.0
lark==1.1.9
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.42.0
locket==1.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
mpi4py==3.1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numba==0.59.1
numpy==1.26.4
nvidia-ammo==0.7.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nvtx==0.2.10
onnx==1.16.0
onnx-graphsurgeon==0.5.2
onnxruntime==1.16.3
optimum==1.19.0
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
panel==1.4.1
param==2.1.0
parso==0.8.4
partd==1.4.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
plotly==5.21.0
polygraphy==0.49.9
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pycparser==2.22
pyct==0.5.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
pygraphviz==1.13
pylibcugraph-cu12==24.4.0
pylibraft-cu12==24.4.0
pynndescent==0.5.12
pynvjitlink-cu12==0.1.14
pynvml==11.4.1
pyparsing==3.1.2
pyproj==3.6.1
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pyvis==0.3.2
pyviz_comms==3.0.2
PyYAML==6.0.1
pyzmq==26.0.1
raft-dask-cu12==24.4.0
rapids-dask-dependency==24.4.1
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rmm-cu12==24.4.0
rpds-py==0.18.0
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.2.0
shapely==2.0.4
simpervisor==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
StrEnum==0.4.15
sympy==1.12
tblib==3.0.0
tenacity==8.2.3
tensorrt==9.3.0.post12.dev1
tensorrt-bindings==9.3.0.post12.dev1
tensorrt-libs==9.3.0.post12.dev1
terminado==0.18.1
thinc==8.2.3
threadpoolctl==3.4.0
tifffile==2024.4.18
tinycss2==1.2.1
tokenizers==0.15.2
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.38.2
treelite==4.1.2
triton==2.2.0
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uc-micro-py==1.0.3
ucx-py-cu12==0.37.0
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
xarray==2024.3.0
xxhash==3.4.1
xyzservices==2024.4.0
yarl==1.9.4
zict==3.0.0
zipp==3.18.1
